+++ extensions/dicom-pdf/package.json
@@ -20,6 +20,7 @@
   ],
   "scripts": {
     "dev": "webpack --config .webpack/webpack.dev.js --watch --debug --output-pathinfo",
+    "dev:dicom-pdf": "yarn run dev",
     "build": "webpack --config .webpack/webpack.prod.js",
     "build:package": "yarn run build",
     "prepublishOnly": "yarn run build",
