+++ extensions/vtk/package.json
@@ -21,6 +21,7 @@
   },
   "scripts": {
     "dev": "webpack --config .webpack/webpack.dev.js --watch --debug --output-pathinfo",
+    "dev:dicom-vtk": "yarn run dev",
     "build": "webpack --config .webpack/webpack.prod.js",
     "build:package": "yarn run build",
     "prepublishOnly": "yarn run build",
