+++ platform/viewer/.webpack/webpack.common.js
