+++ platform/viewer/.webpack/webpack.prod.js
